Revert "changed int to long"
This reverts commit 58eb149d2a7ebaf1fdf927a23f1bad1a2f81422e.

+++ src/main/java/com/BilBay/bilbay/models/Address.java
@@ -19,7 +19,7 @@ public class Address {
     private String address;
 
     @Column(name = "telephone")
-    private long telephone;
+    private int telephone;
 
     @Column (name = "city")
     private String city;
